Add conftests for pipelines, ensembles
